requirements.txt
# Bibliotecas principais
streamlit>=1.24.0
pandas>=2.0.3
numpy>=1.24.3
scipy>=1.10.1
matplotlib>=3.7.2

# Visualização
plotly>=5.15.0
altair>=5.0.1

# Machine Learning
scikit-learn>=1.3.0

# Manipulação de dados
joblib>=1.3.1
openpyxl>=3.1.2

# Interface
streamlit-extras>=0.3.0
streamlit-plotly-events>=0.0.6
statsmodels>=0.14.4

seaborn>=0.11.0

pyyaml>=6.0

plotly>=5.5.0

openpyxl>=3.0.9

xlrd>=2.0.1
xlwt>=1.3.0

jinja2>=3.0.0
